refactor(generate-g5): extract diff subsystem without behavior drift
